docs: rename python_libraries to bigframes_queries
