+++ bigquery/bigframes/bigframes_queries_test.py
@@ -14,32 +14,32 @@
 
 import pytest
 
-import python_libraries
+import bigframes_queries
 
 
 @pytest.mark.skip(reason="Placeholder project ID 'your-project-id' (b/522845525)")
 def test_query_standard_sql():
-    df = python_libraries.query_standard_sql()
+    df = bigframes_queries.query_standard_sql()
     assert df is not None
 
 
 @pytest.mark.skip(reason="Legacy SQL syntax not supported by BigQuery DataFrames (b/522845525)")
 def test_query_legacy_sql():
-    df = python_libraries.query_legacy_sql()
+    df = bigframes_queries.query_legacy_sql()
     assert df is not None
 
 
 def test_query_bqstorage():
-    pandas_df = python_libraries.query_bqstorage()
+    pandas_df = bigframes_queries.query_bqstorage()
     assert pandas_df is not None
 
 
 def test_query_parameters():
-    df = python_libraries.query_parameters()
+    df = bigframes_queries.query_parameters()
     assert df is not None
 
 
 @pytest.mark.skip(reason="Requires a writable table destination (b/522845525)")
 def test_upload_from_dataframe():
-    bq_df = python_libraries.upload_from_dataframe()
+    bq_df = bigframes_queries.upload_from_dataframe()
     assert bq_df is not None
